requirements.txt
numpy
pygame
sounddevice
torch
scipy
transformers
openai
langchain
python-dotenv
